a separated playlist does not work
unfortunately it is not the same as i thought.
i can not create a playlist specially for douban.fm since
mpd will not remove song from this playlist anyway which
means there is no chance for me to detect the total number
of songs in the playlist is below a threshhold.
so before i can find any better solution, just make use
of the default playlist which is dynamic.